requirements.txt
python-dotenv==0.17.0
python-telegram-bot==13.11